+++ api/build.gradle
@@ -42,8 +42,8 @@ publishing {
 }
 
 dependencies {
-    compileOnly "org.spigotmc:spigot-api:1.16.4-R0.1-SNAPSHOT"
-    compileOnly "dev.jaims.mcutils:bukkit:2.1.3"
+    compileOnly 'com.destroystokyo.paper:paper-api:1.16.4-R0.1-SNAPSHOT'
+    compileOnly "dev.jaims.mcutils:bukkit:2.2.4"
     compileOnly "me.mattstudios:mf-msg-bukkit:2.0.1"
 }
