+++ bukkit/build.gradle
@@ -1,13 +1,13 @@
 // add the dependencies
 dependencies {
-    compileOnly "org.spigotmc:spigot-api:1.16.4-R0.1-SNAPSHOT"
+    compileOnly 'com.destroystokyo.paper:paper-api:1.16.4-R0.1-SNAPSHOT'
     compileOnly "me.clip:placeholderapi:2.10.9"
     compileOnly "com.github.MilkBowl:VaultAPI:1.7"
 
     implementation project(":api")
+    implementation "dev.jaims.mcutils:bukkit:2.2.4"
 
     pdm "com.github.jkcclemens:khttp:39f76b4186"
     pdm "me.mattstudios:mf-msg-bukkit:2.0.1"
     pdm "me.mattstudios:triumph-config:1.0.1-SNAPSHOT"
-    pdm "dev.jaims.mcutils:bukkit:2.1.3"
 }
