+++ bukkit/src/main/kotlin/dev/jaims/moducore/bukkit/event/listener/PlayerChatListener.kt
@@ -24,8 +24,8 @@
 
 package dev.jaims.moducore.bukkit.event.listener
 
-import dev.jaims.mcutils.bukkit.async
-import dev.jaims.mcutils.bukkit.send
+import dev.jaims.mcutils.bukkit.util.async
+import dev.jaims.mcutils.bukkit.util.send
 import dev.jaims.moducore.api.event.ModuCoreAsyncChatEvent
 import dev.jaims.moducore.bukkit.ModuCore
 import dev.jaims.moducore.bukkit.config.Config
