+++ bukkit/src/main/kotlin/dev/jaims/moducore/bukkit/event/listener/PlayerInteractListener.kt
@@ -24,7 +24,7 @@
 
 package dev.jaims.moducore.bukkit.event.listener
 
-import dev.jaims.mcutils.bukkit.colorize
+import dev.jaims.mcutils.bukkit.util.colorize
 import dev.jaims.moducore.api.event.ModuCoreSignCommandEvent
 import dev.jaims.moducore.bukkit.ModuCore
 import dev.jaims.moducore.bukkit.config.Modules
