+++ bukkit/src/main/kotlin/dev/jaims/moducore/bukkit/external/VaultEconomyProvider.kt
@@ -1,7 +1,32 @@
+/*
+ * This file is a part of ModuCore, licensed under the MIT License.
+ *
+ * Copyright (c) 2020 James Harrell
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
 package dev.jaims.moducore.bukkit.external
 
 import dev.jaims.mcutils.common.getUUID
 import dev.jaims.moducore.bukkit.ModuCore
+import dev.jaims.moducore.bukkit.api.manager.DefaultEconomyManager
 import dev.jaims.moducore.bukkit.config.Config
 import dev.jaims.moducore.bukkit.config.Modules
 import dev.jaims.moducore.bukkit.util.decimalFormat
@@ -14,14 +39,14 @@ import org.bukkit.plugin.ServicePriority
 class VaultEconomyProvider(private val plugin: ModuCore) : AbstractEconomy()
 {
 
-    private val economyManager = plugin.api.economyManager
+    private val economyManager: DefaultEconomyManager by lazy { plugin.api.economyManager }
 
     companion object
     {
 
         val FAILURE = EconomyResponse(0.0, 0.0, EconomyResponse.ResponseType.FAILURE, null)
         val BANKS_NOT_SUPPORTED =
-                EconomyResponse(0.0, 0.0, EconomyResponse.ResponseType.NOT_IMPLEMENTED, "Banks are not supported.")
+            EconomyResponse(0.0, 0.0, EconomyResponse.ResponseType.NOT_IMPLEMENTED, "Banks are not supported.")
 
     }
 
@@ -31,7 +56,8 @@ class VaultEconomyProvider(private val plugin: ModuCore) : AbstractEconomy()
         plugin.server.servicesManager.register(Economy::class.java, this, plugin, ServicePriority.Highest)
     }
 
-    fun unregister() {
+    fun unregister()
+    {
         plugin.server.servicesManager.unregister(Economy::class.java, this)
     }
 
@@ -86,10 +112,10 @@ class VaultEconomyProvider(private val plugin: ModuCore) : AbstractEconomy()
      */
     override fun currencyNameSingular(): String = plugin.api.fileManager.config.getProperty(Config.CURRENCY_SINGULAR)
 
-    override fun hasAccount(playerName: String): Boolean
-    {
-        TODO("Not yet implemented")
-    }
+    /**
+     * Everyone has an account. you get one when you join
+     */
+    override fun hasAccount(playerName: String): Boolean = true
 
     /**
      * Return the balance or -1 if its negative
