+++ bukkit/src/main/kotlin/dev/jaims/moducore/bukkit/util/CommandSenderExtension.kt
@@ -24,7 +24,7 @@
 
 package dev.jaims.moducore.bukkit.util
 
-import dev.jaims.mcutils.bukkit.send
+import dev.jaims.mcutils.bukkit.util.send
 import dev.jaims.moducore.bukkit.ModuCore
 import dev.jaims.moducore.bukkit.config.Lang
 import org.bukkit.command.CommandSender
